+++ system/src/Grav/Common/Grav.php
@@ -257,7 +257,7 @@ class Grav extends Container
             $this['session']->close();
         }
 
-        if ($this['uri']->isExternal($route)) {
+        if ($uri->isExternal($route)) {
             $url = $route;
         } else {
             $url = rtrim($uri->rootUrl(), '/') .'/'. trim($route, '/');
@@ -277,7 +277,6 @@ class Grav extends Container
     {
         /** @var Language $language */
         $language = $this['language'];
-        $config = $this['config'];
 
         if ($language->enabled()) {
             return $this->redirect($language->getLanguage() . $route, $code);
@@ -413,7 +412,7 @@ class Grav extends Container
     }
 
     /**
-     * This attempts to fine media, other files, and download them
+     * This attempts to find media, other files, and download them
      * @param $page
      * @param $path
      */